新增LCModel_KK和LCModel_ACC换道模型，以及UI优化等（目前版本环形边界多车道场景仍存在bug）

+++ trasim_simplified/core/frame/circle_lane.py
@@ -10,7 +10,7 @@ from trasim_simplified.msg.trasimWarning import TrasimWarning
 
 
 class LaneCircle(LaneAbstract):
-    def __init__(self, lane_length: int):
+    def __init__(self, lane_length: float):
         super().__init__(lane_length)
         self.is_circle = True
